converted primitives over to using the registered dclass handlers

+++ internal/mbean/object.go
@@ -62,7 +62,7 @@ func createContainerReference(env *jnigi.Env, value string, elementTypePath stri
 		json.Unmarshal(byteStr, &arr)
 		return createGenericContainerReference(env, containerTypePath, arr)
 	case LONG:
-		arr := make([]int64, 0)
+		arr := make([]int, 0)
 		json.Unmarshal(byteStr, &arr)
 		return createGenericContainerReference(env, containerTypePath, arr)
 	case BOOLEAN:
